requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>3.9.1
djangorestframework-jwt==1.11.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
pyjwt>=1.7.1
